ENH: (check-valid tasks) Honor the `--skip-cache` flag and provide consistent output
Running the tasks with

```bash
grunt check-valid --skip-cache=true
```

Will now delete the cache files. Subsequent calls to this task without that flag will use a new cache file.
This is useful when a configuration file has changed and we want to destroy the cache that was created
when using the previous configuration, since the files didn't change, the tasks would assume they are OK.

TODO: Store the configuration as well as part of the cache, that way if the config change the task will destroy
the cache automatically and create a new one for the next executions without actually having to use the `--skip-cache` flag. Fix #5

Also in this commit. Normalized the output of the checkvalid tasks. Fix #6